// ui/src/lib/api-client/model/registerDto.ts
/**
 * NestJS API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegisterDto { 
    /**
     * User email address
     */
    email: string;
    /**
     * User password
     */
    password: string;
    /**
     * User first name
     */
    firstName: string;
    /**
     * User last name
     */
    lastName: string;
}